Finished Handling User Input
