+++ frontend/src/components/GetStation.tsx
@@ -1,0 +1,43 @@
+import React, { useState } from "react";
+
+// interface StationProps {
+//   stationID: number;
+// }
+// export type stationID = {
+//   stationID: number;
+// };
+
+type StationProps = {
+  onSubmit: (stationID: string) => void;
+};
+
+const GetStation = ({ onSubmit }: StationProps) => {
+  const [stationID, setStationID] = useState<string>("");
+
+  const handleSubmit = (e: React.FormEvent<HTMLFormElement>) => {
+    e.preventDefault();
+    if (typeof stationID === "string") {
+      onSubmit(stationID);
+    }
+  };
+
+  const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+    const newValue = e.target.value;
+    setStationID(newValue);
+  };
+
+  return (
+    <form className="input" onSubmit={handleSubmit}>
+      <input
+        type="text"
+        value={stationID || ""}
+        placeholder="Enter the Station ID"
+        onChange={handleChange}
+      />
+
+      <button type="submit">Lookup</button>
+    </form>
+  );
+};
+
+export default GetStation;
