+++ frontend/src/station.ts
@@ -1,0 +1,3 @@
+export interface StationID {
+  stationid: number;
+}
